Remove legacy app routes and controllers
